services/tasks/requirements.txt
flask
flask-cors==3.0.6
flask-restful
flask-testing
flask-mongoengine
marshmallow
flask-marshmallow
pymongo
marshmallow_mongoengine
mongoengine
pytest
gunicorn==19.8.1
coverage==4.5.1
flake8===3.5.0
flask-debugtoolbar==0.10.1
flask-bcrypt==0.7.1
pyjwt==1.6.4
